feat: implement signout logic
