+++ src/components/App/App.jsx
@@ -1,8 +1,8 @@
 import './App.css';
 import React from 'react';
 import { useState } from 'react';
-import { Navigate, Route, Routes } from 'react-router-dom';
-import MainScreen from '../Test';
+import { Navigate, Route, Routes, useNavigate } from 'react-router-dom';
+import MainScreen from '../MainScreen/MainScreen';
 import LoginScreen from '../LoginScreen/LoginScreen';
 import ProtectedRoute from '../ProtectedRoute/ProtectedRoute';
 
@@ -11,9 +11,17 @@ function App() {
   const [isLoggedIn, setIsLoggedIn] = useState(false);
   const authorized = localStorage.getItem('authorized');
 
+  const push = useNavigate();
+
   const handleLogin = () => {
     localStorage.setItem('authorized', true);
-    setIsLoggedIn(true)
+    setIsLoggedIn(true);
+  }
+
+  const handleSignout = () => {
+    localStorage.removeItem('authorized');
+    setIsLoggedIn(false);
+    push('/')
   }
 
   return (
@@ -23,7 +31,7 @@ function App() {
           <Route path='/'
             element={
               <ProtectedRoute isLoggedIn={authorized}>
-                <MainScreen />
+                <MainScreen handleSignout={handleSignout}/>
               </ProtectedRoute>
             }
           />
