+++ src/components/Header/Header.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Header({ handleSignout }) {
+  return (
+    <div className="header">
+      <button onClick={handleSignout}>Signout</button>
+    </div>
+  )
+}
